Fix hyperv panic
Move SyncIC related code into vmbus driver. So that hv_hv.c contains
only hypercall related code so it is ready to move inside kernel.
