+++ sys/dev/hyperv/vmbus/hv_hv.c
@@ -36,12 +36,12 @@ __FBSDID("$FreeBSD$");
 #include <sys/malloc.h>
 #include <sys/pcpu.h>
 #include <sys/timetc.h>
+#include <sys/kernel.h>
 #include <machine/bus.h>
 #include <machine/md_var.h>
 #include <vm/vm.h>
 #include <vm/vm_param.h>
 #include <vm/pmap.h>
-
 
 #include "hv_vmbus_priv.h"
 
@@ -50,13 +50,7 @@ __FBSDID("$FreeBSD$");
 
 static u_int hv_get_timecount(struct timecounter *tc);
 
-/**
- * Globals
- */
-hv_vmbus_context hv_vmbus_g_context = {
-	.syn_ic_initialized = FALSE,
-	.hypercall_page = NULL,
-};
+static void*		hypercall_page = NULL;
 
 static struct timecounter hv_timecounter = {
 	hv_get_timecount, 0, ~0u, HV_NANOSECONDS_PER_SEC/100, "Hyper-V", HV_NANOSECONDS_PER_SEC/100
@@ -73,7 +67,7 @@ hv_get_timecount(struct timecounter *tc)
  * @brief Query the cpuid for presence of windows hypervisor
  */
 int
-hv_vmbus_query_hypervisor_presence(void) 
+hv_vmbus_query_hypervisor_presence(void)
 {
 	if (vm_guest != VM_GUEST_HV)
 		return (0);
@@ -120,7 +114,6 @@ hv_vmbus_do_hypercall(uint64_t control, void* input, void* output)
 	uint64_t hv_status = 0;
 	uint64_t input_address = (input) ? hv_get_phys_addr(input) : 0;
 	uint64_t output_address = (output) ? hv_get_phys_addr(output) : 0;
-	volatile void* hypercall_page = hv_vmbus_g_context.hypercall_page;
 
 	__asm__ __volatile__ ("mov %0, %%r8" : : "r" (output_address): "r8");
 	__asm__ __volatile__ ("call *%3" : "=a"(hv_status):
@@ -138,7 +131,6 @@ hv_vmbus_do_hypercall(uint64_t control, void* input, void* output)
 	uint64_t output_address = (output) ? hv_get_phys_addr(output) : 0;
 	uint32_t output_address_high = output_address >> 32;
 	uint32_t output_address_low = output_address & 0xFFFFFFFF;
-	volatile void* hypercall_page = hv_vmbus_g_context.hypercall_page;
 
 	__asm__ __volatile__ ("call *%8" : "=d"(hv_status_high),
 				"=a"(hv_status_low) : "d" (control_high),
@@ -156,25 +148,15 @@ hv_vmbus_do_hypercall(uint64_t control, void* input, void* output)
  *  This routine must be called
  *  before any other routines in here are called
  */
-int
-hv_vmbus_init(void) 
+static int
+hv_vmbus_init(void* context)
 {
 	int					max_leaf;
 	hv_vmbus_x64_msr_hypercall_contents	hypercall_msr;
-	void* 					virt_addr = 0;
-
-	memset(
-	    hv_vmbus_g_context.syn_ic_event_page,
-	    0,
-	    sizeof(hv_vmbus_handle) * MAXCPU);
-
-	memset(
-	    hv_vmbus_g_context.syn_ic_msg_page,
-	    0,
-	    sizeof(hv_vmbus_handle) * MAXCPU);
+	void*					virt_addr;
 
 	if (vm_guest != VM_GUEST_HV)
-	    goto cleanup;
+		return (ENOTSUP);
 
 	max_leaf = hv_vmbus_get_hypervisor_version();
 
@@ -183,7 +165,6 @@ hv_vmbus_init(void)
 	 */
 	uint64_t os_guest_info = HV_FREEBSD_GUEST_ID;
 	wrmsr(HV_X64_MSR_GUEST_OS_ID, os_guest_info);
-	hv_vmbus_g_context.guest_id = os_guest_info;
 
 	/*
 	 * See if the hypercall page is already set
@@ -205,43 +186,36 @@ hv_vmbus_init(void)
 	if (!hypercall_msr.u.enable)
 	    goto cleanup;
 
-	hv_vmbus_g_context.hypercall_page = virt_addr;
+	hypercall_page = virt_addr;
 
 	tc_init(&hv_timecounter); /* register virtual timecount */
 
-	hv_et_init();
-	
 	return (0);
 
 	cleanup:
-	if (virt_addr != NULL) {
-	    if (hypercall_msr.u.enable) {
+	if (hypercall_msr.u.enable) {
 		hypercall_msr.as_uint64_t = 0;
-		wrmsr(HV_X64_MSR_HYPERCALL,
-					hypercall_msr.as_uint64_t);
-	    }
-
-	    free(virt_addr, M_DEVBUF);
+		wrmsr(HV_X64_MSR_HYPERCALL, hypercall_msr.as_uint64_t);
 	}
+
+	free(virt_addr, M_DEVBUF);
 	return (ENOTSUP);
 }
 
 /**
  * @brief Cleanup routine, called normally during driver unloading or exiting
  */
-void
-hv_vmbus_cleanup(void) 
+static void
+hv_vmbus_cleanup(void* context)
 {
 	hv_vmbus_x64_msr_hypercall_contents hypercall_msr;
 
-	if (hv_vmbus_g_context.guest_id == HV_FREEBSD_GUEST_ID) {
-	    if (hv_vmbus_g_context.hypercall_page != NULL) {
+	if (hypercall_page != NULL) {
 		hypercall_msr.as_uint64_t = 0;
 		wrmsr(HV_X64_MSR_HYPERCALL,
 					hypercall_msr.as_uint64_t);
-		free(hv_vmbus_g_context.hypercall_page, M_DEVBUF);
-		hv_vmbus_g_context.hypercall_page = NULL;
-	    }
+		free(hypercall_page, M_DEVBUF);
+		hypercall_page = NULL;
 	}
 }
 
@@ -307,119 +281,5 @@ hv_vmbus_signal_event(void *con_id)
 	return (status);
 }
 
-/**
- * @brief hv_vmbus_synic_init
- */
-void
-hv_vmbus_synic_init(void *arg)
-
-{
-	int			cpu;
-	uint64_t		hv_vcpu_index;
-	hv_vmbus_synic_simp	simp;
-	hv_vmbus_synic_siefp	siefp;
-	hv_vmbus_synic_scontrol sctrl;
-	hv_vmbus_synic_sint	shared_sint;
-	uint64_t		version;
-	hv_setup_args* 		setup_args = (hv_setup_args *)arg;
-
-	cpu = PCPU_GET(cpuid);
-
-	if (hv_vmbus_g_context.hypercall_page == NULL)
-	    return;
-
-	/*
-	 * TODO: Check the version
-	 */
-	version = rdmsr(HV_X64_MSR_SVERSION);
-	
-	hv_vmbus_g_context.syn_ic_msg_page[cpu] =
-	    setup_args->page_buffers[2 * cpu];
-	hv_vmbus_g_context.syn_ic_event_page[cpu] =
-	    setup_args->page_buffers[2 * cpu + 1];
-
-	/*
-	 * Setup the Synic's message page
-	 */
-
-	simp.as_uint64_t = rdmsr(HV_X64_MSR_SIMP);
-	simp.u.simp_enabled = 1;
-	simp.u.base_simp_gpa = ((hv_get_phys_addr(
-	    hv_vmbus_g_context.syn_ic_msg_page[cpu])) >> PAGE_SHIFT);
-
-	wrmsr(HV_X64_MSR_SIMP, simp.as_uint64_t);
-
-	/*
-	 * Setup the Synic's event page
-	 */
-	siefp.as_uint64_t = rdmsr(HV_X64_MSR_SIEFP);
-	siefp.u.siefp_enabled = 1;
-	siefp.u.base_siefp_gpa = ((hv_get_phys_addr(
-	    hv_vmbus_g_context.syn_ic_event_page[cpu])) >> PAGE_SHIFT);
-
-	wrmsr(HV_X64_MSR_SIEFP, siefp.as_uint64_t);
-
-	/*HV_SHARED_SINT_IDT_VECTOR + 0x20; */
-	shared_sint.as_uint64_t = 0;
-	shared_sint.u.vector = setup_args->vector;
-	shared_sint.u.masked = FALSE;
-	shared_sint.u.auto_eoi = TRUE;
-
-	wrmsr(HV_X64_MSR_SINT0 + HV_VMBUS_MESSAGE_SINT,
-	    shared_sint.as_uint64_t);
-
-	/* Enable the global synic bit */
-	sctrl.as_uint64_t = rdmsr(HV_X64_MSR_SCONTROL);
-	sctrl.u.enable = 1;
-
-	wrmsr(HV_X64_MSR_SCONTROL, sctrl.as_uint64_t);
-
-	hv_vmbus_g_context.syn_ic_initialized = TRUE;
-
-	/*
-	 * Set up the cpuid mapping from Hyper-V to FreeBSD.
-	 * The array is indexed using FreeBSD cpuid.
-	 */
-	hv_vcpu_index = rdmsr(HV_X64_MSR_VP_INDEX);
-	hv_vmbus_g_context.hv_vcpu_index[cpu] = (uint32_t)hv_vcpu_index;
-
-	return;
-}
-
-/**
- * @brief Cleanup routine for hv_vmbus_synic_init()
- */
-void hv_vmbus_synic_cleanup(void *arg)
-{
-	hv_vmbus_synic_sint	shared_sint;
-	hv_vmbus_synic_simp	simp;
-	hv_vmbus_synic_siefp	siefp;
-
-	if (!hv_vmbus_g_context.syn_ic_initialized)
-	    return;
-
-	shared_sint.as_uint64_t = rdmsr(
-	    HV_X64_MSR_SINT0 + HV_VMBUS_MESSAGE_SINT);
-
-	shared_sint.u.masked = 1;
-
-	/*
-	 * Disable the interrupt
-	 */
-	wrmsr(
-	    HV_X64_MSR_SINT0 + HV_VMBUS_MESSAGE_SINT,
-	    shared_sint.as_uint64_t);
-
-	simp.as_uint64_t = rdmsr(HV_X64_MSR_SIMP);
-	simp.u.simp_enabled = 0;
-	simp.u.base_simp_gpa = 0;
-
-	wrmsr(HV_X64_MSR_SIMP, simp.as_uint64_t);
-
-	siefp.as_uint64_t = rdmsr(HV_X64_MSR_SIEFP);
-	siefp.u.siefp_enabled = 0;
-	siefp.u.base_siefp_gpa = 0;
-
-	wrmsr(HV_X64_MSR_SIEFP, siefp.as_uint64_t);
-}
-
+SYSINIT(hv_init, SI_SUB_HYPERVISOR, SI_ORDER_ANY, hv_vmbus_init, NULL);
+SYSUNINIT(hv_cleanup, SI_SUB_HYPERVISOR, SI_ORDER_ANY, hv_vmbus_cleanup, NULL);
