+++ sys/dev/hyperv/vmbus/hv_vmbus_drv_freebsd.c
@@ -65,10 +65,16 @@ __FBSDID("$FreeBSD$");
 #include "acpi_if.h"
 
 static device_t vmbus_devp;
-static int vmbus_inited;
 static hv_setup_args setup_args; /* only CPU 0 supported at this time */
 
 static char *vmbus_ids[] = { "VMBUS", NULL };
+
+/**
+ * Globals
+ */
+hv_vmbus_context hv_vmbus_g_context = {
+	.syn_ic_initialized = FALSE,
+};
 
 /**
  * @brief Software interrupt thread routine to handle channel messages from
@@ -394,6 +400,129 @@ vmbus_vector_alloc(void)
 }
 
 /**
+ * @brief hv_vmbus_synic_init
+ */
+void
+hv_vmbus_synic_init(void *arg)
+
+{
+	int			cpu;
+	uint64_t		hv_vcpu_index;
+	hv_vmbus_synic_simp	simp;
+	hv_vmbus_synic_siefp	siefp;
+	hv_vmbus_synic_scontrol sctrl;
+	hv_vmbus_synic_sint	shared_sint;
+	uint64_t		version;
+	hv_setup_args* 		setup_args = (hv_setup_args *)arg;
+
+	cpu = PCPU_GET(cpuid);
+
+	memset(
+	    hv_vmbus_g_context.syn_ic_event_page,
+	    0,
+	    sizeof(hv_vmbus_handle) * MAXCPU);
+
+	memset(
+	    hv_vmbus_g_context.syn_ic_msg_page,
+	    0,
+	    sizeof(hv_vmbus_handle) * MAXCPU);
+
+	/*
+	 * TODO: Check the version
+	 */
+	version = rdmsr(HV_X64_MSR_SVERSION);
+
+	hv_vmbus_g_context.syn_ic_msg_page[cpu] =
+	    setup_args->page_buffers[2 * cpu];
+	hv_vmbus_g_context.syn_ic_event_page[cpu] =
+	    setup_args->page_buffers[2 * cpu + 1];
+
+	/*
+	 * Setup the Synic's message page
+	 */
+
+	simp.as_uint64_t = rdmsr(HV_X64_MSR_SIMP);
+	simp.u.simp_enabled = 1;
+	simp.u.base_simp_gpa = ((hv_get_phys_addr(
+	    hv_vmbus_g_context.syn_ic_msg_page[cpu])) >> PAGE_SHIFT);
+
+	wrmsr(HV_X64_MSR_SIMP, simp.as_uint64_t);
+
+	/*
+	 * Setup the Synic's event page
+	 */
+	siefp.as_uint64_t = rdmsr(HV_X64_MSR_SIEFP);
+	siefp.u.siefp_enabled = 1;
+	siefp.u.base_siefp_gpa = ((hv_get_phys_addr(
+	    hv_vmbus_g_context.syn_ic_event_page[cpu])) >> PAGE_SHIFT);
+
+	wrmsr(HV_X64_MSR_SIEFP, siefp.as_uint64_t);
+
+	/*HV_SHARED_SINT_IDT_VECTOR + 0x20; */
+	shared_sint.as_uint64_t = 0;
+	shared_sint.u.vector = setup_args->vector;
+	shared_sint.u.masked = FALSE;
+	shared_sint.u.auto_eoi = TRUE;
+
+	wrmsr(HV_X64_MSR_SINT0 + HV_VMBUS_MESSAGE_SINT,
+	    shared_sint.as_uint64_t);
+
+	/* Enable the global synic bit */
+	sctrl.as_uint64_t = rdmsr(HV_X64_MSR_SCONTROL);
+	sctrl.u.enable = 1;
+
+	wrmsr(HV_X64_MSR_SCONTROL, sctrl.as_uint64_t);
+
+	hv_vmbus_g_context.syn_ic_initialized = TRUE;
+
+	/*
+	 * Set up the cpuid mapping from Hyper-V to FreeBSD.
+	 * The array is indexed using FreeBSD cpuid.
+	 */
+	hv_vcpu_index = rdmsr(HV_X64_MSR_VP_INDEX);
+	hv_vmbus_g_context.hv_vcpu_index[cpu] = (uint32_t)hv_vcpu_index;
+
+	return;
+}
+
+/**
+ * @brief Cleanup routine for hv_vmbus_synic_init()
+ */
+void hv_vmbus_synic_cleanup(void *arg)
+{
+	hv_vmbus_synic_sint	shared_sint;
+	hv_vmbus_synic_simp	simp;
+	hv_vmbus_synic_siefp	siefp;
+
+	if (!hv_vmbus_g_context.syn_ic_initialized)
+	    return;
+
+	shared_sint.as_uint64_t = rdmsr(
+	    HV_X64_MSR_SINT0 + HV_VMBUS_MESSAGE_SINT);
+
+	shared_sint.u.masked = 1;
+
+	/*
+	 * Disable the interrupt
+	 */
+	wrmsr(
+	    HV_X64_MSR_SINT0 + HV_VMBUS_MESSAGE_SINT,
+	    shared_sint.as_uint64_t);
+
+	simp.as_uint64_t = rdmsr(HV_X64_MSR_SIMP);
+	simp.u.simp_enabled = 0;
+	simp.u.base_simp_gpa = 0;
+
+	wrmsr(HV_X64_MSR_SIMP, simp.as_uint64_t);
+
+	siefp.as_uint64_t = rdmsr(HV_X64_MSR_SIEFP);
+	siefp.u.siefp_enabled = 0;
+	siefp.u.base_siefp_gpa = 0;
+
+	wrmsr(HV_X64_MSR_SIEFP, siefp.as_uint64_t);
+}
+
+/**
  * @brief Restore the IDT slot to rsvd.
  */
 static void
@@ -450,19 +579,6 @@ vmbus_bus_init(void)
 	char buf[MAXCOMLEN + 1];
 	cpuset_t cpu_mask;
 
-	if (vmbus_inited)
-		return (0);
-
-	vmbus_inited = 1;
-
-	ret = hv_vmbus_init();
-
-	if (ret) {
-		if(bootverbose)
-			printf("Error VMBUS: Hypervisor Initialization Failed!\n");
-		return (ret);
-	}
-
 	/*
 	 * Find a free IDT slot for vmbus callback.
 	 */
@@ -472,7 +588,7 @@ vmbus_bus_init(void)
 		if(bootverbose)
 			printf("Error VMBUS: Cannot find free IDT slot for "
 			    "vmbus callback!\n");
-		goto cleanup;
+		return (ENOTSUP);
 	}
 
 	if(bootverbose)
@@ -553,15 +669,6 @@ vmbus_bus_init(void)
 
 	smp_rendezvous(NULL, hv_vmbus_synic_init, NULL, &setup_args);
 
-	/*
-	 * Connect to VMBus in the root partition
-	 */
-	ret = hv_vmbus_connect();
-
-	if (ret != 0)
-		goto cleanup1;
-
-	hv_vmbus_request_channel_offers();
 	return (ret);
 
 	cleanup1:
@@ -580,13 +687,10 @@ vmbus_bus_init(void)
 			taskqueue_free(hv_vmbus_g_context.hv_event_queue[j]);
 		if (hv_vmbus_g_context.msg_swintr[j] != NULL)
 			swi_remove(hv_vmbus_g_context.msg_swintr[j]);
-		hv_vmbus_g_context.hv_msg_intr_event[j] = NULL;	
+		hv_vmbus_g_context.hv_msg_intr_event[j] = NULL;
 	}
 
 	vmbus_vector_free(hv_vmbus_g_context.hv_cb_vector);
-
-	cleanup:
-	hv_vmbus_cleanup();
 
 	return (ret);
 }
@@ -598,18 +702,26 @@ vmbus_attach(device_t dev)
 		device_printf(dev, "VMBUS: attach dev: %p\n", dev);
 	vmbus_devp = dev;
 
-	vmbus_bus_init();
-
 	return (0);
 }
 
-static void
-vmbus_init(void)
+static int
+vmbus_bus_connect(void* context)
 {
-	if (vm_guest != VM_GUEST_HV)
-		return;
+	int ret;
 
-	vmbus_bus_init();
+	hv_et_init();
+
+	/*
+	 * Connect to VMBus in the root partition
+	 */
+	ret = hv_vmbus_connect();
+
+	if (ret != 0)
+		return (ret);
+
+	hv_vmbus_request_channel_offers();
+	return (ret);
 }
 
 static void
@@ -627,15 +739,13 @@ vmbus_bus_exit(void)
 			free(setup_args.page_buffers[i], M_DEVBUF);
 	}
 
-	hv_vmbus_cleanup();
-
 	/* remove swi */
 	CPU_FOREACH(i) {
 		if (hv_vmbus_g_context.hv_event_queue[i] != NULL)
 			taskqueue_free(hv_vmbus_g_context.hv_event_queue[i]);
 		if (hv_vmbus_g_context.msg_swintr[i] != NULL)
 			swi_remove(hv_vmbus_g_context.msg_swintr[i]);
-		hv_vmbus_g_context.hv_msg_intr_event[i] = NULL;	
+		hv_vmbus_g_context.hv_msg_intr_event[i] = NULL;
 	}
 
 	vmbus_vector_free(hv_vmbus_g_context.hv_cb_vector);
@@ -643,44 +753,25 @@ vmbus_bus_exit(void)
 	return;
 }
 
-static void
-vmbus_exit(void)
-{
-	vmbus_bus_exit();
-}
-
 static int
 vmbus_detach(device_t dev)
 {
-	vmbus_exit();
+	vmbus_bus_exit();
 	return (0);
-}
-
-static void
-vmbus_mod_load(void)
-{
-	if(bootverbose)
-		printf("VMBUS: load\n");
-}
-
-static void
-vmbus_mod_unload(void)
-{
-	if(bootverbose)
-		printf("VMBUS: unload\n");
 }
 
 static int
 vmbus_modevent(module_t mod, int what, void *arg)
 {
+	if (vm_guest != VM_GUEST_HV)
+		return (ENXIO);
+
 	switch (what) {
 
 	case MOD_LOAD:
-		vmbus_init();
-		vmbus_mod_load();
+		vmbus_bus_init();
 		break;
 	case MOD_UNLOAD:
-		vmbus_mod_unload();
 		break;
 	}
 
@@ -707,10 +798,11 @@ static device_method_t vmbus_methods[] = {
 static char driver_name[] = "vmbus";
 static driver_t vmbus_driver = { driver_name, vmbus_methods,0, };
 
-
 devclass_t vmbus_devclass;
 
-DRIVER_MODULE(vmbus, acpi, vmbus_driver, vmbus_devclass, vmbus_modevent, 0);
+EARLY_DRIVER_MODULE(vmbus, acpi, vmbus_driver, vmbus_devclass, vmbus_modevent, 0,
+	BUS_PASS_BUS);
 MODULE_DEPEND(vmbus, acpi, 1, 1, 1);
 MODULE_VERSION(vmbus, 1);
 
+SYSINIT(vmbus_connect, SI_SUB_VFS, SI_ORDER_ANY, vmbus_bus_connect, NULL);
